backend/requirements.txt
flask
flask-cors
gitpython 